# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Movie AutoTranslate API",
    "version": "0.1.0"
  },
  "paths": {
    "/healthz": {
      "get": {
        "tags": [
          "health"
        ],
        "summary": "Healthz",
        "operationId": "healthz_healthz_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Healthz Healthz Get"
                }
              }
            }
          }
        }
      }
    },
    "/v1/upload/init": {
      "post": {
        "tags": [
          "uploads"
        ],
        "summary": "Init Upload",
        "operationId": "init_upload_v1_upload_init_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UploadInitRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadInitResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/upload/complete": {
      "post": {
        "tags": [
          "uploads"
        ],
        "summary": "Complete Upload",
        "operationId": "complete_upload_v1_upload_complete_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UploadCompleteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Complete Upload V1 Upload Complete Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/jobs": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "List Jobs",
        "operationId": "list_jobs_v1_jobs_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Pagesize"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/jobs/translate": {
      "post": {
        "tags": [
          "jobs"
        ],
        "summary": "Create Translation Job",
        "operationId": "create_translation_job_v1_jobs_translate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/jobs/{job_id}": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "Get Job",
        "operationId": "get_job_v1_jobs__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "jobs"
        ],
        "summary": "Cancel Job",
        "operationId": "cancel_job_v1_jobs__job_id__delete",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/jobs/{job_id}/retry": {
      "post": {
        "tags": [
          "jobs"
        ],
        "summary": "Retry Job",
        "operationId": "retry_job_v1_jobs__job_id__retry_post",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobRetryRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/assets/{asset_id}": {
      "get": {
        "tags": [
          "assets"
        ],
        "summary": "Get Asset",
        "operationId": "get_asset_v1_assets__asset_id__get",
        "parameters": [
          {
            "name": "asset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Asset Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssetResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/assets/{asset_id}/hls/master.m3u8": {
      "get": {
        "tags": [
          "assets"
        ],
        "summary": "Get Asset Master Playlist",
        "operationId": "get_asset_master_playlist_v1_assets__asset_id__hls_master_m3u8_get",
        "parameters": [
          {
            "name": "asset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Asset Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Get Asset Master Playlist V1 Assets  Asset Id  Hls Master M3U8 Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/metrics": {
      "get": {
        "summary": "Metrics",
        "operationId": "metrics_metrics_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AssetResponse": {
        "properties": {
          "userId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Userid"
          },
          "srcLang": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "en",
                  "es",
                  "fr",
                  "de"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Srclang"
          },
          "targetLangs": {
            "items": {
              "type": "string",
              "enum": [
                "en",
                "es",
                "fr",
                "de"
              ]
            },
            "type": "array",
            "title": "Targetlangs"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "assetId": {
            "type": "string",
            "title": "Assetid"
          },
          "storageKeys": {
            "$ref": "#/components/schemas/StorageKeys"
          },
          "durationSec": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Durationsec"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "title": "Updatedat"
          },
          "outputs": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Outputs"
          }
        },
        "type": "object",
        "required": [
          "id",
          "assetId",
          "storageKeys",
          "createdAt",
          "updatedAt"
        ],
        "title": "AssetResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "JobCreateRequest": {
        "properties": {
          "assetId": {
            "type": "string",
            "title": "Assetid"
          },
          "targetLangs": {
            "items": {
              "type": "string",
              "enum": [
                "en",
                "es",
                "fr",
                "de"
              ]
            },
            "type": "array",
            "title": "Targetlangs"
          },
          "presets": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Presets"
          },
          "resumeFrom": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/JobStage"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "assetId",
          "targetLangs"
        ],
        "title": "JobCreateRequest"
      },
      "JobListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/JobResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "pageSize": {
            "type": "integer",
            "title": "Pagesize"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "pageSize"
        ],
        "title": "JobListResponse"
      },
      "JobResponse": {
        "properties": {
          "jobId": {
            "type": "string",
            "title": "Jobid"
          },
          "assetId": {
            "type": "string",
            "title": "Assetid"
          },
          "stage": {
            "$ref": "#/components/schemas/JobStage"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus"
          },
          "progress": {
            "type": "number",
            "title": "Progress"
          },
          "startedAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Startedat"
          },
          "endedAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Endedat"
          },
          "failedStage": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/JobStage"
              },
              {
                "type": "null"
              }
            ]
          },
          "errorMessage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Errormessage"
          },
          "targetLangs": {
            "items": {
              "type": "string",
              "enum": [
                "en",
                "es",
                "fr",
                "de"
              ]
            },
            "type": "array",
            "title": "Targetlangs"
          },
          "presets": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Presets"
          },
          "stageHistory": {
            "additionalProperties": {
              "type": "object"
            },
            "type": "object",
            "title": "Stagehistory"
          },
          "logsKey": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Logskey"
          }
        },
        "type": "object",
        "required": [
          "jobId",
          "assetId",
          "stage",
          "status",
          "progress"
        ],
        "title": "JobResponse"
      },
      "JobRetryRequest": {
        "properties": {
          "resumeFrom": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/JobStage"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "JobRetryRequest"
      },
      "JobStage": {
        "type": "string",
        "enum": [
          "INGESTED",
          "ASR",
          "TRANSLATE",
          "TTS",
          "ALIGN/MIX",
          "PACKAGE",
          "PUBLISHED",
          "DONE"
        ],
        "title": "JobStage"
      },
      "JobStatus": {
        "type": "string",
        "enum": [
          "PENDING",
          "RUNNING",
          "SUCCESS",
          "FAILED",
          "CANCELLED"
        ],
        "title": "JobStatus"
      },
      "StorageKeys": {
        "properties": {
          "raw": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Raw"
          },
          "processed": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Processed"
          },
          "public": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Public"
          }
        },
        "type": "object",
        "title": "StorageKeys"
      },
      "UploadCompleteRequest": {
        "properties": {
          "assetId": {
            "type": "string",
            "title": "Assetid"
          },
          "uploadId": {
            "type": "string",
            "title": "Uploadid"
          },
          "etags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Etags"
          },
          "srcLang": {
            "type": "string",
            "title": "Srclang"
          },
          "targetLangs": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Targetlangs"
          }
        },
        "type": "object",
        "required": [
          "assetId",
          "uploadId",
          "etags",
          "srcLang",
          "targetLangs"
        ],
        "title": "UploadCompleteRequest"
      },
      "UploadInitRequest": {
        "properties": {
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "contentType": {
            "type": "string",
            "title": "Contenttype"
          },
          "size": {
            "type": "integer",
            "title": "Size"
          }
        },
        "type": "object",
        "required": [
          "filename",
          "contentType",
          "size"
        ],
        "title": "UploadInitRequest"
      },
      "UploadInitResponse": {
        "properties": {
          "assetId": {
            "type": "string",
            "title": "Assetid"
          },
          "uploadId": {
            "type": "string",
            "title": "Uploadid"
          },
          "partSize": {
            "type": "integer",
            "title": "Partsize"
          },
          "parts": {
            "items": {
              "$ref": "#/components/schemas/UploadPart"
            },
            "type": "array",
            "title": "Parts"
          }
        },
        "type": "object",
        "required": [
          "assetId",
          "uploadId",
          "partSize",
          "parts"
        ],
        "title": "UploadInitResponse"
      },
      "UploadPart": {
        "properties": {
          "partNumber": {
            "type": "integer",
            "title": "Partnumber"
          },
          "uploadUrl": {
            "type": "string",
            "title": "Uploadurl"
          }
        },
        "type": "object",
        "required": [
          "partNumber",
          "uploadUrl"
        ],
        "title": "UploadPart"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}